+++ e2e/test/scenarios/question/notebook-data-source.cy.spec.ts
@@ -28,6 +28,8 @@ import {
   visitModel,
   visitQuestion,
   visualize,
+  openOrdersTable,
+  queryBuilderMain,
 } from "e2e/support/helpers";
 import { checkNotNull } from "metabase/lib/types";
 
@@ -391,6 +393,80 @@ describe("scenarios > notebook > data source", { tags: "@OSS" }, () => {
       entityPickerModalTab("Saved questions").should("not.exist");
     });
   });
+});
+
+describe("issue 34350", { tags: "@external" }, () => {
+  beforeEach(() => {
+    restore("postgres-12");
+    cy.signInAsAdmin();
+  });
+
+  it("works after changing question's source table to a one from a different database (metabase#34350)", () => {
+    openOrdersTable({ mode: "notebook" });
+    openDataSelector();
+    entityPickerModal().within(() => {
+      cy.findByText("QA Postgres12").click();
+      cy.findByText("Orders").click();
+    });
+
+    visualize();
+
+    queryBuilderMain()
+      .findByText("There was a problem with your question")
+      .should("not.exist");
+    cy.findAllByTestId("cell-data").should("contain", "37.65");
+  });
+});
+
+describe("issue 28106", () => {
+  beforeEach(() => {
+    const dialect = "postgres";
+
+    resetTestTable({ type: dialect, table: "many_schemas" });
+    restore(`${dialect}-writable`);
+    cy.signInAsAdmin();
+
+    resyncDatabase({ dbId: WRITABLE_DB_ID });
+  });
+
+  it(
+    "should not jump to the top of schema list when scrolling (metabase#28106)",
+    { tags: "@external" },
+    () => {
+      startNewQuestion();
+      entityPickerModal().within(() => {
+        entityPickerModalTab("Tables").click();
+        cy.findByText("Writable Postgres12").click();
+
+        entityPickerModalLevel(1)
+          .findByTestId("scroll-container")
+          .as("schemasList");
+
+        // The list is virtualized and the scrollbar height changes during scrolling (metabase#44966)
+        // that's why we need to scroll and wait multiple times.
+        // Test is flaky because of this - that's why there are 3 attempts.
+        for (let i = 0; i < 3; ++i) {
+          cy.get("@schemasList").scrollTo("bottom");
+          cy.wait(100);
+        }
+
+        // assert scrolling worked and the last item is visible
+        entityPickerModalItem(1, "Public").should("be.visible");
+
+        // simulate scrolling up using mouse wheel 3 times
+        for (let i = 0; i < 3; ++i) {
+          cy.get("@schemasList").realMouseWheel({ deltaY: -100 });
+          cy.wait(100);
+        }
+
+        // assert first item does not exist - this means the list has not been scrolled to the top
+        cy.findByText("Domestic").should("not.exist");
+        cy.get("@schemasList").should(([$element]) => {
+          expect($element.scrollTop).to.be.greaterThan(0);
+        });
+      });
+    },
+  );
 });
 
 function moveToCollection(collection: string) {