+++ frontend/src/metabase/models/components/ModelDetailPage/ModelInfoSidePanel/ModelInfoSidePanel.tsx
@@ -40,6 +40,7 @@ function ModelInfoSidePanel({ model, mainTable, onChangeDescription }: Props) {
           }
           isOptional
           isMultiline
+          isMarkdown
           isDisabled={!canWrite}
           aria-label={t`Description`}
           onChange={onChangeDescription}