+++ frontend/src/metabase/models/components/ModelDetailPage/ModelUsageDetails/ModelUsageDetails.styled.tsx
@@ -10,7 +10,6 @@ export const CardListItem = styled(Link)`
   display: flex;
   flex-direction: row;
   align-items: center;
-
   width: 100%;
   border-radius: 8px;
   padding: 1rem 0.5rem;