+++ frontend/src/metabase/query_builder/components/view/sidebars/SummarizeSidebar/AddAggregationButton/AddAggregationButton.styled.tsx
@@ -4,7 +4,6 @@ import Button from "metabase/core/components/Button";
 
 export const AddAggregationButtonRoot = styled(Button)`
   padding: 0.625rem;
-
   color: var(--mb-color-summarize);
   background-color: var(--mb-color-bg-light);