+++ frontend/src/metabase/querying/components/FilterModal/FilterModal.tsx
@@ -1,7 +1,8 @@
 import { useMemo, useState } from "react";
 
+import CS from "metabase/css/core/index.css";
 import { FilterContent } from "metabase/querying/components/FilterContent";
-import { Modal } from "metabase/ui";
+import { Flex, Modal } from "metabase/ui";
 import * as Lib from "metabase-lib";
 
 import { ModalBody, ModalFooter, ModalHeader } from "./FilterModal.styled";
@@ -121,7 +122,9 @@ export function FilterModal({
       <Modal.Content>
         <ModalHeader p="lg">
           <Modal.Title>{getModalTitle(groupItems)}</Modal.Title>
-          <FilterContent.Header value={searchText} onChange={handleSearch} />
+          <Flex mx="md" justify="end" className={CS.flex1}>
+            <FilterContent.Header value={searchText} onChange={handleSearch} />
+          </Flex>
           <Modal.CloseButton />
         </ModalHeader>
         <ModalBody p={0}>