+++ frontend/test/__support__/server-mocks/field.ts
@@ -41,14 +41,15 @@ export function setupFieldsValuesEndpoints(
   fieldsValues.forEach(fieldValues => setupFieldValuesEndpoints(fieldValues));
 }
 
-export function setupFieldSearchValuesEndpoints(
+export function setupFieldSearchValuesEndpoint(
   fieldId: FieldId,
+  searchFieldId: FieldId,
   searchValue: string,
   result: FieldValue[] = [],
 ) {
   fetchMock.get(
     {
-      url: `path:/api/field/${fieldId}/search/${fieldId}`,
+      url: `path:/api/field/${fieldId}/search/${searchFieldId}`,
       query: {
         value: searchValue,
         limit: 100, // corresponds to MAX_SEARCH_RESULTS in FieldValuesWidget
@@ -57,5 +58,6 @@ export function setupFieldSearchValuesEndpoints(
     {
       body: result,
     },
+    { overwriteRoutes: false },
   );
 }