+++ frontend/test/__support__/ui.tsx
@@ -15,18 +15,18 @@ import { Router, useRouterHistory } from "react-router";
 import { routerReducer, routerMiddleware } from "react-router-redux";
 import _ from "underscore";
 
-import { AppInitializeController } from "embedding-sdk/components/private/AppInitializeController";
-import { SdkThemeProvider } from "embedding-sdk/components/private/SdkThemeProvider";
+import {
+  MetabaseProviderInternal,
+  type MetabaseProviderProps,
+} from "embedding-sdk/components/public/MetabaseProvider";
 import { sdkReducers } from "embedding-sdk/store";
 import type { SdkStoreState } from "embedding-sdk/store/types";
 import { createMockSdkState } from "embedding-sdk/test/mocks/state";
-import type { SDKConfig } from "embedding-sdk/types";
 import { Api } from "metabase/api";
 import { UndoListing } from "metabase/containers/UndoListing";
 import { baseStyle } from "metabase/css/core/base.styled";
 import { mainReducers } from "metabase/reducers-main";
 import { publicReducers } from "metabase/reducers-public";
-import { EmotionCacheProvider } from "metabase/styled-components/components/EmotionCacheProvider";
 import { ThemeProvider } from "metabase/ui";
 import type { State } from "metabase-types/store";
 import { createMockState } from "metabase-types/store/mocks";
@@ -51,7 +51,7 @@ export interface RenderWithProvidersOptions {
   withDND?: boolean;
   withUndos?: boolean;
   customReducers?: ReducerObject;
-  sdkConfig?: SDKConfig | null;
+  sdkProviderProps?: Partial<MetabaseProviderProps> | null;
   theme?: MantineThemeOverride;
 }
 
@@ -71,7 +71,7 @@ export function renderWithProviders(
     withDND = false,
     withUndos = false,
     customReducers,
-    sdkConfig = null,
+    sdkProviderProps = null,
     theme,
     ...options
   }: RenderWithProvidersOptions = {},
@@ -130,7 +130,11 @@ export function renderWithProviders(
   const wrapper = (props: any) => {
     if (mode === "sdk") {
       return (
-        <SdkWrapper {...props} config={sdkConfig} store={store} theme={theme} />
+        <MetabaseProviderInternal
+          {...props}
+          {...sdkProviderProps}
+          store={store}
+        />
       );
     }
 
@@ -199,31 +203,6 @@ export function TestWrapper({
           {withUndos && <UndoListing />}
         </ThemeProvider>
       </MaybeDNDProvider>
-    </Provider>
-  );
-}
-
-function SdkWrapper({
-  config,
-  children,
-  store,
-}: {
-  config: SDKConfig;
-  children: React.ReactElement;
-  store: any;
-  history?: History;
-  withRouter: boolean;
-  withDND: boolean;
-}) {
-  return (
-    <Provider store={store}>
-      <EmotionCacheProvider>
-        <SdkThemeProvider>
-          <AppInitializeController config={config}>
-            {children}
-          </AppInitializeController>
-        </SdkThemeProvider>
-      </EmotionCacheProvider>
     </Provider>
   );
 }