+++ release/src/linked-issues.ts
@@ -1,6 +1,6 @@
 export function getLinkedIssues(body: string) {
   const matches = body.match(
-    /(close(s|d)?|fixe?(s|d)?|resolve(s|d)?) (#|https?:\/\/github.com\/.+\/issues\/)(\d+)/gi,
+    /(close(s|d)?|fixe?(s|d)?|resolve(s|d)?)(:?) (#|https?:\/\/github.com\/.+\/issues\/)(\d+)/gi,
   );
 
   if (matches) {
@@ -14,9 +14,10 @@ export function getLinkedIssues(body: string) {
 }
 
 export function getPRsFromCommitMessage(message: string) {
-  const result = [ ...message.matchAll(/\(#(\d+)\)/g) ];
+  const firstLine = message.split('\n\n')[0];
+  const result = [ ...firstLine.matchAll(/\(#(\d+)\)/g) ];
   if (!result.length) {
-    console.log('no pr found in commit message', message);
+    console.log('No pr found in commit message', message);
     return null;
   }